requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
sentence-transformers>=2.2.2
numpy>=1.24.3
python-dotenv==1.0.0
torch>=2.0.0
transformers>=4.21.0
gunicorn==21.2.0